Moved zombsole code to a zombsole package, preparing for setuptools.
