+++ zombsole/maps/maze_for_safehouse
