+++ zombsole/maps/village_for_evacuation
