+++ zombsole/players/sniper.py
@@ -1,7 +1,7 @@
 # coding: utf-8
-from things import Player, Zombie
-from utils import closest
-from weapons import Rifle
+from zombsole.things import Player, Zombie
+from zombsole.utils import closest
+from zombsole.weapons import Rifle
 
 
 class Sniper(Player):
